worker prereqs i.e no func passed
